+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## current
+
+- Made `Allocator` trait public.
+- Added `Renderer::with_allocator` which can take any type implementing `Allocator`.
+
+### Breaking
+
+- Renderer is now generic over the `Allocator` trait.
+- Remove `dynamic-rendering` feature.
+- Functions that used to take `vk::RenderPass` or `DynamicRendering` now take `RenderMode` instead.
+
 ## 0.12.0
 
 ### Breaking